♻️ refactor: 타입 파일명 변경 #257
