+++ client/src/types/auth.type.ts
